refactor: бин NamedParameterJdbcTemplate добавлен в dao, добавлены классы UserMapper и FilmMapper

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDbStorage.java
@@ -2,9 +2,10 @@ package ru.yandex.practicum.filmorate.storage.film;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Primary;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.PreparedStatementCreator;
 import org.springframework.jdbc.core.ResultSetExtractor;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
+import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
+import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
@@ -12,13 +13,7 @@ import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-import java.util.Collection;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
+import java.util.*;
 import java.util.stream.Collectors;
 
 @Primary
@@ -26,17 +21,15 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class FilmDbStorage implements FilmStorage {
 
-    private final JdbcTemplate jdbcTemplate;
-
+    private final NamedParameterJdbcTemplate customJdbcTemplate;
     private final ResultSetExtractor<Map<Film, List<Genre>>> filmExtractor;
 
     @Override
     public Collection<Film> findAll() {
-        String sqlGetAllFilms = "\nSELECT *\n" +
-                "FROM films f \n" +
+        String sqlGetAllFilms = "SELECT * FROM films f \n" +
                 "LEFT JOIN genre_films gf ON f.film_id = gf.film_id \n" +
                 "LEFT JOIN genre g ON gf.genre_id = g.genre_id";
-        var mapGenre = jdbcTemplate.query(sqlGetAllFilms, filmExtractor);
+        var mapGenre = customJdbcTemplate.query(sqlGetAllFilms, filmExtractor);
         for (var film : mapGenre.entrySet()) {
             film.getKey().setGenres(film.getValue());
         }
@@ -46,17 +39,23 @@ public class FilmDbStorage implements FilmStorage {
     @Override
     public Film add(Film film) {
         String sqlAddFilm =
-                "INSERT INTO films (FILM_NAME,  DESCRIPTION, RELEASE_DATE, DURATION_MINUTE, MPA_id) " +
-                        "VALUES (?,?,?,?,?);";
+                "INSERT INTO films (film_name,  description, release_date, duration_minute, MPA_id) " +
+                        "VALUES (:film_name, :description, :release_date, :duration_minute, :MPA_id);";
+
+        Map<String, Object> params = new HashMap<>();
+        params.put("film_name", film.getName());
+        params.put("description", film.getDescription());
+        params.put("release_date", film.getReleaseDate());
+        params.put("duration_minute", film.getDuration().getSeconds());
+        params.put("MPA_id", film.getMpa().getId());
 
         KeyHolder keyHolder = new GeneratedKeyHolder();
 
-        PreparedStatementCreator preparedStatementCreator = con -> makeStatement(con, film, sqlAddFilm);
+        SqlParameterSource paramSource = new MapSqlParameterSource(params);
 
-        jdbcTemplate.update(preparedStatementCreator, keyHolder);
+        customJdbcTemplate.update(sqlAddFilm, paramSource, keyHolder);
 
         film.setId(keyHolder.getKey().longValue());
-
         return film;
 
     }
@@ -64,16 +63,19 @@ public class FilmDbStorage implements FilmStorage {
     @Override
     public Film update(Film film) {
 
-        String sql = "UPDATE films SET FILM_NAME=?, DESCRIPTION=?, RELEASE_DATE=?, DURATION_MINUTE=?, MPA_id=?" +
-                "WHERE film_id =?;";
+        String sql = "UPDATE films SET film_name = :film_name, description = :description, release_date = :release_date, " +
+                "duration_minute = :duration_minute, MPA_id = :MPA_id " +
+                "WHERE film_id =:film_id;";
 
-        PreparedStatementCreator preparedStatementCreator = con -> {
-            var stmt = makeStatement(con, film, sql);
-            stmt.setString(6, String.valueOf(film.getId()));
-            return stmt;
-        };
+        Map<String, Object> params = new HashMap<>();
+        params.put("film_name", film.getName());
+        params.put("description", film.getDescription());
+        params.put("release_date", film.getReleaseDate());
+        params.put("duration_minute", film.getDuration().getSeconds());
+        params.put("MPA_id", film.getMpa().getId());
+        params.put("film_id", film.getId());
 
-        if (jdbcTemplate.update(preparedStatementCreator) == 0)
+        if (customJdbcTemplate.update(sql, params) == 0)
             throw new ValidationException("Фильма не существует");
 
         return film;
@@ -81,12 +83,13 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public Film getFilm(long id) {
-        String sqlGetFilm = "\nSELECT * \n" +
-                "FROM films f\n" +
+        String sqlGetFilm = "SELECT * FROM films f \n" +
                 "LEFT JOIN genre_films gf ON f.film_id = gf.film_id \n" +
                 "LEFT JOIN genre g ON gf.genre_id = g.genre_id \n" +
-                "WHERE f.film_id = ?";
-        var mapGenre = jdbcTemplate.query(sqlGetFilm, filmExtractor, id);
+                "WHERE f.film_id = :film_id";
+        var params = Collections.singletonMap("film_id", id);
+        var mapGenre = customJdbcTemplate.query(sqlGetFilm, params, filmExtractor);
+
         for (var film : mapGenre.entrySet()) {
             film.getKey().setGenres(film.getValue());
             return film.getKey();
@@ -96,14 +99,20 @@ public class FilmDbStorage implements FilmStorage {
 
     @Override
     public void addLike(long filmId, long userId) {
-        String sql = "MERGE INTO likes_film VALUES (?, ?)";
-        jdbcTemplate.update(sql, filmId, userId);
+        String sql = "MERGE INTO likes_film (film_id, user_id) VALUES (:film_id, :user_id)";
+        Map<String, Object> params = new HashMap<>();
+        params.put("film_id", filmId);
+        params.put("user_id", userId);
+        customJdbcTemplate.update(sql, params);
     }
 
     @Override
     public void removeLike(long userId, long filmId) {
-        String sql = "DELETE FROM likes_film WHERE user_id = ? AND film_id = ?";
-        if (jdbcTemplate.update(sql, userId, filmId) == 0)
+        String sql = "DELETE FROM likes_film WHERE user_id = :user_id AND film_id = :film_id";
+        Map<String, Object> params = new HashMap<>();
+        params.put("film_id", filmId);
+        params.put("user_id", userId);
+        if (customJdbcTemplate.update(sql, params) == 0)
             throw new ValidationException("Введён несуществующий id");
     }
 
@@ -114,11 +123,13 @@ public class FilmDbStorage implements FilmStorage {
                 "(SELECT film_id FROM likes_film \n" +
                 "GROUP BY film_id \n" +
                 "ORDER by COUNT(user_id) DESC \n" +
-                "LIMIT ?) popular_film \n" +
+                "LIMIT :size) popular_film \n" +
                 "LEFT JOIN films f ON f.film_id = popular_film.film_id \n" +
                 "LEFT JOIN genre_films gf ON f.film_id = gf.film_id \n" +
                 "LEFT JOIN genre g ON gf.genre_id = g.genre_id";
-        var mapGenre = jdbcTemplate.query(sqlGetPopularFilms, filmExtractor, size);
+
+        var param = Collections.singletonMap("size", size);
+        var mapGenre = customJdbcTemplate.query(sqlGetPopularFilms, param, filmExtractor);
         for (var film : mapGenre.entrySet()) {
             film.getKey().setGenres(film.getValue());
         }
@@ -126,17 +137,6 @@ public class FilmDbStorage implements FilmStorage {
         if (listPopular.isEmpty())
             listPopular = findAll().stream().limit(size).collect(Collectors.toSet());
         return listPopular;
-    }
-
-    private PreparedStatement makeStatement(Connection con, Film film, String sql) throws SQLException {
-        PreparedStatement stmt = con.prepareStatement(sql, new String[]{"film_id"});
-        stmt.setString(1, film.getName());
-        stmt.setString(2, film.getDescription());
-        if (film.getMpa() != null)
-            stmt.setString(5, String.valueOf(film.getMpa().getId()));
-        stmt.setString(3, film.getReleaseDate().toString());
-        stmt.setString(4, String.valueOf(film.getDuration().getSeconds()));
-        return stmt;
     }
 
 }
